+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - 添加报错提示映射
 - 添加 Vue 相关 Loader
 - 添加 Vue 的 ESLint 检查
+- 将 `Webpack.config.js` 的开发与生产模式配置区分